Fixed a bug that causes unnecessary unwrapping
We called `as_json` too early in many cases, among other things this has the
effect of casing the encoder to miss the caching envelope for the comments, so
the comment fragments weren't being re-used. After this fix, the reuse
performance improved significantly, but it appears to have made the raw encoding
slower. Need to look closer to see why.